Adiciona plugin para sincronização dos planos de ação do transferegov
